fixed date sort function
